backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.135.2",
    "google-genai>=1.69.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.42.0",
    "yt-dlp>=2026.3.17",
]
